[Mobile] check if a path exists before deleting it task : 91526
[git-p4: depot-paths = "//depot/4eDimension/main/QMobile/QMobileUI/": change = 225935]

+++ Sources/Application/Services/ApplicationCrashManager.swift
@@ -246,12 +246,16 @@ extension ApplicationCrashManager {
 
     fileprivate func deleteCrashFile(pathCrash: Path, zipPath: Path) {
         do {
-            try pathCrash.deleteFile()
+            if pathCrash.exists {
+                try pathCrash.deleteFile()
+            }
         } catch {
             logger.warning("Failed to delete crash file \(error.localizedDescription)")
         }
         do {
-            try zipPath.deleteFile()
+            if zipPath.exists {
+                try zipPath.deleteFile()
+            }
         } catch {
             logger.warning("Failed to delete zipped crash file \(error.localizedDescription)")
         }
